leverage native dir attribute for RTL and handle fallback
