+++ packages/components/pie-webc-core/src/mixins/rtl/rtlMixin.ts
@@ -1,74 +1,80 @@
 /* eslint-disable max-classes-per-file */
-import { LitElement } from 'lit';
-import { property } from 'lit/decorators/property.js';
+import { LitElement, isServer } from 'lit';
 
-// According to TS, "A mixin class must have a constructor with a single rest parameter of type 'any[]'."
+/**
+ * A type representing a constructor of any class.
+ * @typedef {new (...args: any[]) => T} Constructor
+ * @template T
+ */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type Constructor<T> = new (...args: any[]) => T;
 
-type htmlDirAttribute = 'ltr' | 'rtl' | 'auto';
-
 /**
- * Any component property interface that implements RTL should extend this interface. See the ModalProps interface for an example of this.
+ * An interface representing the structure of RTL related class.
+ * @interface
  */
-export interface RTLComponentProps {
-    dir: htmlDirAttribute;
-}
-
-// This is just used by the dynamically constructed class below and does not need to be imported or referenced anywhere else
 declare class _RTLInterface {
-    dir: htmlDirAttribute;
+    /** A boolean indicating whether the text direction is right-to-left. */
     isRTL: boolean;
 }
 
 /**
- * This RTL mixin is used with Lit Web components to add programmatic Right-to-Left (RTL) support.
- * It is only required if your component either:
- * - Needs RTL awareness in its TypeScript logic.
- * - Its CSS requires a [dir='rtl'] attribute to be present.
+ * A mixin to extend LitElement with right-to-left (RTL) text direction handling.
+ * This mixin adds a read-only `isRTL` property to the class it's applied to,
+ * allowing you to easily determine the text direction within your component.
  *
- * By default, components will infer the `dir` property from the document's root `dir` attribute.
- * If needed, it's possible to override specific components direction by setting the `dir` property
- * value. The `dir` property value is reflected in the DOM, making it queryable. The `isRTL`
- * internal property returns true or false depending on whether or not the `dir` property is `rtl` or not.
+ * @function
+ * @param {Constructor<LitElement>} superClass - The LitElement class to extend with RTL functionality.
+ * @returns {Constructor<_RTLInterface> & T} - A new class extending both LitElement and _RTLInterface.
  *
- * **Note:** If you provide a `dir` property manually, we strongly suggest you use either `ltr` or `rtl` as the value and avoid using `auto`.
+ * @example
+ * ```typescript
+ * import { LitElement, html } from 'lit';
+ * import { RtlMixin } from '@justeattakeaway/pie-webc-core';
  *
- * ---
- * **SSR Usage:** There is no document.documentElement to infer the direction from on the server. For this reason, if no dir is set on the component, it will default to 'auto`
- * and infer from the root element on the client-side.
+ * class MyElement extends RtlMixin(LitElement) {
+ *   render() {
+ *     return html`Text direction is ${this.isRTL ? 'right-to-left' : 'left-to-right'}`;
+ *   }
+ * }
  *
- * If you require the component to have a `dir` set during SSR, you need need to manually set the dir property on the component.
- *  * @example
- * // Manually provide a dir property
- * <pie-component dir="rtl"></pie-component>
+ * customElements.define('my-element', MyElement);
+ * ```
  */
-
 export const RtlMixin =
     <T extends Constructor<LitElement>>(superClass: T) => {
+        /**
+       * Class representing a LitElement with RTL handling.
+       * @extends {LitElement}
+       * @implements {_RTLInterface}
+       */
         class RTLElement extends superClass implements _RTLInterface {
-            // Initialized with a default value. Updated later if on the client-side.
-            @property({ type: String, reflect: true })
-            public dir: htmlDirAttribute = 'auto';
-
-            connectedCallback (): void {
-                super.connectedCallback();
-                if (this.dir === 'auto') {
-                    this.dir = document.documentElement.dir as htmlDirAttribute || 'auto';
-                }
-            }
-
             /**
-             * Returns true if the element is in Right to Left mode.
-             * If a dir attribute is not explicitly set on the web component
-             * then it falls back to the nearest parent with a dir attribute set.
+             * A getter to determine whether the text direction is right-to-left (RTL).
+             * If the `dir` property is present on the component, it will be used to determine the text direction.
+             * If running on the client-side (not SSR) and the `dir` property is not present, the text direction will be inferred
+             * from the document's root element. This inference is not available during SSR.
+             * In all other cases, it will return `false`, indicating a left-to-right (LTR) text direction.
              *
-             * A dir attribute being set will result in a reactive property.
-             * If the component falls back to a parent dir attribute then the value
-             * will not be reactive and is only computed once
+             * @returns {boolean} - Returns `true` if the text direction is RTL, otherwise `false`.
              */
             get isRTL (): boolean {
-                return this.dir === 'rtl';
+                if (this.dir) {
+                    return this.dir === 'rtl';
+                }
+
+                // This check is necessary as document is undefined during SSR
+                if (isServer) {
+                    // Inference from the document's root element is not available during SSR.
+                    return false;
+                }
+
+                // If running on client-side and `dir` is not present, infer from the document's root element.
+                if (!isServer && !this.dir) {
+                    return document.documentElement.getAttribute('dir') === 'rtl';
+                }
+
+                return false;
             }
         }
